refactor

+++ FileUploadAndValidation/Models/InitiatePaymentRequest.cs
@@ -6,13 +6,13 @@ namespace FileUploadAndValidation.Models
 {
     public class InitiatePaymentRequest
     {
-        public int BusinessId { get; set; }
+        public long BusinessId { get; set; }
 
         public long UserId { get; set; }
 
         public string UserName { get; set; }
 
-        public int ApprovalConfigId { get; set; }
+        public long ApprovalConfigId { get; set; }
 
     }
 }
